+++ src/dataStructures/tree/avlTree/types/index.ts
@@ -2,5 +2,6 @@ import { BinarySearchTreeInterface,
   BinarySearchTreeNodeInterface } from '../../binarySearchTree/types';
 
 export interface AvlTreeInterface<T = unknown> extends BinarySearchTreeInterface<T> {
+  binarySearchTree: BinarySearchTreeInterface<T>;
   balance: (node: BinarySearchTreeNodeInterface<T>) => this;
 }