+++ src/dataStructures/tree/redBlackRree/redBlackTree.ts
@@ -142,7 +142,7 @@ export class RedBlackTree<T = unknown> implements RedBlackTreeInterface<T> {
     currentNode.value.makeBlack();
   }
 
-  private readonly binarySearchTree: BinarySearchTreeInterface<RedBlackTreeNodeInterface<T>>;
+  public readonly binarySearchTree: BinarySearchTreeInterface<RedBlackTreeNodeInterface<T>>;
 
   public setRoot(root: CompleteRedBlackTreeNode<T> | null): this {
     this.binarySearchTree.setRoot(root);